refactor: change name nextlogo
